FG1 - Copy/workspace/empty/a1/new-ui/backup/banking-api-full/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
cryptography
requests
